Refactor CSV output handling by consolidating CsvPredictions functionality into FilePredictions and simplifying code structure.
